+++ backend/app/models/help.py
@@ -1,5 +1,5 @@
 """帮助中心模型"""
-from sqlalchemy import Column, Integer, String, Text
+from sqlalchemy import Column, Integer, BigInteger, String, Text
 from app.models.base import Base, TimestampMixin
 
 
@@ -7,7 +7,7 @@ class HelpCategory(Base, TimestampMixin):
     """帮助分类表"""
     __tablename__ = "help_categories"
 
-    id = Column(Integer, primary_key=True, index=True, autoincrement=True)
+    id = Column(BigInteger, primary_key=True, index=True, autoincrement=True)
     name = Column(String(100), nullable=False, comment="分类名称")
     slug = Column(String(100), unique=True, nullable=False, index=True, comment="分类标识")
     description = Column(Text, nullable=True, comment="分类描述")
@@ -20,8 +20,8 @@ class HelpArticle(Base, TimestampMixin):
     """帮助文章表"""
     __tablename__ = "help_articles"
 
-    id = Column(Integer, primary_key=True, index=True, autoincrement=True)
-    category_id = Column(Integer, nullable=False, index=True, comment="分类ID")
+    id = Column(BigInteger, primary_key=True, index=True, autoincrement=True)
+    category_id = Column(BigInteger, nullable=False, index=True, comment="分类ID")
     title = Column(String(200), nullable=False, comment="文章标题")
     slug = Column(String(200), unique=True, nullable=False, index=True, comment="文章标识")
     content = Column(Text, nullable=False, comment="文章内容")
@@ -37,8 +37,8 @@ class FAQ(Base, TimestampMixin):
     """常见问题表"""
     __tablename__ = "faqs"
 
-    id = Column(Integer, primary_key=True, index=True, autoincrement=True)
-    category_id = Column(Integer, nullable=True, index=True, comment="分类ID")
+    id = Column(BigInteger, primary_key=True, index=True, autoincrement=True)
+    category_id = Column(BigInteger, nullable=True, index=True, comment="分类ID")
     question = Column(String(500), nullable=False, comment="问题")
     answer = Column(Text, nullable=False, comment="答案")
     tags = Column(String(200), nullable=True, comment="标签，逗号分隔")
