+++ backend/app/models/ticket.py
@@ -1,5 +1,5 @@
 """客服工单模型"""
-from sqlalchemy import Column, Integer, String, Text, ForeignKey, DateTime
+from sqlalchemy import Column, Integer, BigInteger, String, Text, ForeignKey, DateTime
 from sqlalchemy.orm import relationship
 from app.models.base import Base, TimestampMixin
 
@@ -8,8 +8,8 @@ class Ticket(Base, TimestampMixin):
     """工单表"""
     __tablename__ = "tickets"
 
-    id = Column(Integer, primary_key=True, index=True, autoincrement=True)
-    user_id = Column(Integer, ForeignKey("users.id"), nullable=False, index=True)
+    id = Column(BigInteger, primary_key=True, index=True, autoincrement=True)
+    user_id = Column(BigInteger, ForeignKey("users.id"), nullable=False, index=True)
     ticket_no = Column(String(50), unique=True, nullable=False, index=True, comment="工单编号")
     category = Column(String(50), nullable=False, comment="工单分类")
     subject = Column(String(200), nullable=False, comment="工单主题")
@@ -28,9 +28,9 @@ class TicketReply(Base, TimestampMixin):
     """工单回复表"""
     __tablename__ = "ticket_replies"
 
-    id = Column(Integer, primary_key=True, index=True, autoincrement=True)
-    ticket_id = Column(Integer, ForeignKey("tickets.id"), nullable=False, index=True)
-    user_id = Column(Integer, nullable=False, index=True, comment="回复人ID")
+    id = Column(BigInteger, primary_key=True, index=True, autoincrement=True)
+    ticket_id = Column(BigInteger, ForeignKey("tickets.id"), nullable=False, index=True)
+    user_id = Column(BigInteger, nullable=False, index=True, comment="回复人ID")
     is_staff = Column(Integer, default=0, nullable=False, comment="是否客服: 0用户 1客服")
     content = Column(Text, nullable=False, comment="回复内容")
     attachments = Column(Text, nullable=True, comment="附件URL，逗号分隔")
@@ -42,8 +42,8 @@ class Feedback(Base, TimestampMixin):
     """用户反馈表"""
     __tablename__ = "feedbacks"
 
-    id = Column(Integer, primary_key=True, index=True, autoincrement=True)
-    user_id = Column(Integer, ForeignKey("users.id"), nullable=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True, autoincrement=True)
+    user_id = Column(BigInteger, ForeignKey("users.id"), nullable=True, index=True)
     feedback_type = Column(String(50), nullable=False, comment="反馈类型")
     content = Column(Text, nullable=False, comment="反馈内容")
     contact = Column(String(100), nullable=True, comment="联系方式")
