+++ backend/app/models/work.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Text, ForeignKey, JSON
+from sqlalchemy import Column, Integer, BigInteger, String, Text, ForeignKey, JSON
 from sqlalchemy.orm import relationship
 from app.models.base import Base, TimestampMixin
 
@@ -6,9 +6,9 @@ from app.models.base import Base, TimestampMixin
 class Work(Base, TimestampMixin):
     __tablename__ = "works"
 
-    id = Column(Integer, primary_key=True, index=True, autoincrement=True)
-    user_id = Column(Integer, ForeignKey("users.id"), nullable=False, index=True)
-    task_id = Column(Integer, ForeignKey("tasks.id"), nullable=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True, autoincrement=True)
+    user_id = Column(BigInteger, ForeignKey("users.id"), nullable=False, index=True)
+    task_id = Column(BigInteger, ForeignKey("tasks.id"), nullable=True, index=True)
     work_type = Column(String(20), nullable=False, comment="作品类型: image/video/ad")
     title = Column(String(200), nullable=True)
     file_url = Column(String(500), nullable=False, comment="文件URL")
